Chore: Refactor move_execution even further
