+++ moved/src/move_execution/gas.rs
@@ -1,0 +1,31 @@
+use {
+    crate::genesis::config::GenesisConfig,
+    aptos_gas_meter::{AptosGasMeter, GasAlgebra, StandardGasAlgebra, StandardGasMeter},
+};
+
+pub fn new_gas_meter(
+    genesis_config: &GenesisConfig,
+    gas_limit: u64,
+) -> StandardGasMeter<StandardGasAlgebra> {
+    StandardGasMeter::new(StandardGasAlgebra::new(
+        genesis_config.gas_costs.version,
+        genesis_config.gas_costs.vm.clone(),
+        genesis_config.gas_costs.storage.clone(),
+        false,
+        gas_limit,
+    ))
+}
+
+pub fn total_gas_used<G: AptosGasMeter>(gas_meter: &G, genesis_config: &GenesisConfig) -> u64 {
+    let gas_algebra = gas_meter.algebra();
+    // Note: this sum is overflow safe because it uses saturating addition
+    // by default in the implementation of `GasQuantity`.
+    let total = gas_algebra.execution_gas_used()
+        + gas_algebra.io_gas_used()
+        + gas_algebra.storage_fee_used_in_gas_units();
+    let total: u64 = total.into();
+    // Aptos scales up the input gas limit for some reason,
+    // so we need to reverse that scaling when we return.
+    let scaling_factor: u64 = genesis_config.gas_costs.vm.txn.scaling_factor().into();
+    total / scaling_factor
+}
